SAK-16802 Adjust code in preparation for allowing content size > 2G (change to match KNL-205 API change)
git-svn-id: https://source.sakaiproject.org/svn/metaobj/trunk@65809 66ffb92e-73f9-0310-93c1-f5514f145a0a

+++ metaobj-api/api/src/java/org/sakaiproject/metaobj/shared/mgt/ContentEntityWrapper.java
@@ -52,7 +52,7 @@ public class ContentEntityWrapper implements ContentResource
 		this.reference = reference;
 	}
 
-	public int getContentLength()
+	public long getContentLength()
 	{
 		return base.getContentLength();
 	}
